Adding quick spot to change $CC if desired.
Note that because we're using Ruby's mkmf library, we cannot simply
export a CC variable to the shell and select a specific compiler.
XCODE >= 4.2 uses llvm, not gcc, and you may find there are moments
when you want to force ruby to build the library using one or the other.
I have not found any problems yet, but have followed the community's lead
and symlinked my /usr/bin/gcc -> llvm-gcc-4.2