Finished conditional compilation changes.
Now, if TSK_VERSION_NUM < 3.2.2 the extra fields found in the
TSK_FS_INFO struct will not be added to the array from which the
Ruby attributes are initialized.  The new macros will allow us to add/hide
features found in Sleuthkit 4.0.x

Tests invoking conditionally compiled methods are edited to check
the version number, and test accordingly.